backend/requirements.txt
# Core backend
fastapi
uvicorn[standard]
pydantic
python-dotenv
PyJWT
passlib[bcrypt]
bcrypt
python-multipart
email-validator
motor
httpx
pymongo
pillow

# Data / ML utilities
pandas
scikit-learn
numpy
joblib
moviepy

# Deep learning / CV
torch
tensorflow
mediapipe
protobuf
FER
# dlib is optional and can be installed on the host if needed (building in-situ often fails)
# dlib
opencv-python-headless
opencv-python
moviepy
librosa
whisper

# Enhanced facial analysis dependencies
deepface
insightface
mtcnn
facenet-pytorch
face-recognition
imutils
scipy
matplotlib
seaborn
fer

# Optional dependencies for better face detection
# dlib  # Uncomment if you want to install dlib (requires compilation)
# face-recognition  # Already included above